+++ sourcemapprocessor/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- perf: enhance symbolication process with per stacktrace error caching | @clintonnkemdilim
+
 ## v0.0.12 [beta] - 2025/10/22
 
 - feat: if `app.debug.source_map_uuid` is in the resource attributes, include it in source paths (#113) | @beekhc